Fix error when quitting

+++ main.js
@@ -26,7 +26,9 @@ function createWindow (railsApp, railsAddr) {
     // in an array if your app supports multi windows, this is the time
     // when you should delete the corresponding element.
 
-    railsApp.kill('SIGINT');
+    if (railsApp) {
+      railsApp.kill('SIGINT');
+    }
     mainWindow = null;
   });
 }
